+++ packages/eslint-plugin/src/rules/space-infix-ops.ts
@@ -18,6 +18,8 @@ const UNIONS = ['|', '&'];
 export default createRule<Options, MessageIds>({
   name: 'space-infix-ops',
   meta: {
+    deprecated: true,
+    replacedBy: ['@stylistic/ts/space-infix-ops'],
     type: 'layout',
     docs: {
       description: 'Require spacing around infix operators',