+++ packages/type-utils/src/predicates.ts
@@ -6,24 +6,23 @@ import { isTypeFlagSet } from './typeFlagUtils';
 
 const log = debug('typescript-eslint:eslint-plugin:utils:types');
 
+export interface IsNullableTypeOptions {
+  /**
+   * @deprecated - this flag no longer does anything and will be removed in the next major
+   */
+  isReceiver?: boolean;
+  /**
+   * @deprecated - this flag no longer does anything and will be removed in the next major
+   */
+  allowUndefined?: boolean;
+}
+
 /**
  * Checks if the given type is (or accepts) nullable
  */
 export function isNullableType(
   type: ts.Type,
-  {
-    isReceiver,
-    allowUndefined,
-  }?: { 
-    /**
-     * @deprecated - this flag no longer does anything and will be removed in the next major
-     */
-    isReceiver?: boolean;
-    /**
-     * @deprecated - this flag no longer does anything and will be removed in the next major
-     */
-    allowUndefined?: boolean;
-  },
+  _deprecated?: IsNullableTypeOptions,
 ): boolean {
   return isTypeFlagSet(
     type,